chore(aws-api): move owner-based auth logic out of plugin
